Chore: friends 파일 삭제
