+++ components/FriendManageList.tsx
@@ -4,8 +4,7 @@ import { useRouter } from "next/navigation";
 import { useSession } from "next-auth/react";
 import { startTransition, useState } from "react";
 import { useToast } from "@/contexts/toast/ToastContext";
-import { getFirstFollow } from "@/lib/actions/friend-actions";
-import { deleteFriend } from "@/lib/actions/friends-action";
+import { getFirstFollow, deleteFriend } from "@/lib/actions/friend-actions";
 import { FriendProfile } from "@/types/common/profile";
 import EmptyState from "./EmptyList";
 import { Txt } from "./atoms";
