+++ lib/actions/friend-actions.ts
@@ -228,3 +228,22 @@ export const postFriendbyId = async (soldierNum: number, userId: number) => {
     };
   }
 };
+
+/**
+ * 친구 삭제
+ * @usage 친구 목록
+ * @param followId
+ * @returns 성공시 success: true 와 지워진 data, 실패 시 success: false
+ */
+export const deleteFriend = async (followId: number) => {
+  try {
+    await prisma.follow.delete({ where: { followId } });
+    return {
+      success: true,
+    };
+  } catch (error) {
+    return {
+      success: false,
+    };
+  }
+};
